+++ src/api/graphql.ts
@@ -1,0 +1,15 @@
+export const RENDER_WIDGET_QUERY = `
+  query renderWidget ($user: UserIdInput, $engagementMedium: UserEngagementMedium, $widgetType: WidgetType, $locale: RSLocale) {
+    renderWidget(user: $user, engagementMedium: $engagementMedium, widgetType: $widgetType, locale: $locale) {
+      template
+      user {
+        id
+        accountId
+      }
+      jsOptions
+      widgetConfig {
+        values
+      }
+    }
+  }
+`;